+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - [Menu] - from Overlays/Menu to Forms/Actions/Menu
 - renaming InformationDisplay to DataDisplay
 - [MapControls] - from presets to Forms/Actions/MapControl/presets/MapControls
+- [IconMarkers] - from presets to Geospatial/IconMarker/presets/IconMarkers
 
 ## [v2.131.0] - 09-10-2025